更新demo
Former-commit-id: c14294955e445116623774051a09c58d13c313bb

+++ src/ZegoExpressExample/SoundLevel/src/main/java/im/zego/soundlevelandspectrum/widget/SpectrumView.java
@@ -88,7 +88,7 @@ public class SpectrumView extends View {
     int[] color = new int[64];
     int maxLine = 64;
     float stopTmp;
-    double[] frequencySpectrums = new double[64];
+    float[] frequencySpectrums = new float[64];
 
     @Override
     protected void onDraw(Canvas canvas) {
@@ -131,7 +131,7 @@ public class SpectrumView extends View {
     }
 
 
-    public void updateFrequencySpectrum(double[] frequencySpectrumList) {
+    public void updateFrequencySpectrum(float[] frequencySpectrumList) {
         frequencySpectrums = frequencySpectrumList;
 
     }
